// REMP/src/types/Property.ts
export interface PropertyData {
    title: string;
    propertyType: number;
    saleCategory: number;
    street: string;
    city: string;
    state: string;
    postcode: number;
    longitude: number;
    latitude: number;
    price: number;
    bedrooms: number;
    bathrooms: number;
    garages: number;
    floorArea: number;
  }